Migrate merge sort
- Move implementation to header file
- Add missing curly braces
- Update README
- Delete old merge_sort.cpp
